Revert addition of auto mode, heat_cool is more appropriate:
CLIMATE_MODE_HEAT_COOL: The climate device is set to heat/cool to reach the target temperature.
CLIMATE_MODE_AUTO: The climate device is adjusting the temperature dynamically. For example, the target temperature can be adjusted based on a schedule, or learned behavior. The target temperature can't be adjusted when in this mode.

Use ESPHome ClimateMode and ClimateAction enumerations directly in s21 control code. HVAC unit reports actions in heat_cool mode in a way that didn't line up with the DaikinClimateMode

Remove mode parameter from load_setpoint, it's always passed the same value and accessible from the function