Climate refactor fix
Publish state as soon as it's received. I was wrong about Home Assistant
handling of UI versus component state. Previous code relied on
incidental changes in state (like current temperature or action) to
publish an update to HA. When there was no change in state the UI would
lag until the command took effect in the unit and that change was
published. You could see this when switching between HEAT_COOL and HEAT
with setpoints resulting in the same action. Now we optimistically
report the desired state and publish a correction later if it doesn't
take.

Remove max_demand setting from demand sensor and provide a configuration
example to do mapping to percent with a standard filter.